+++ app/src/main/java/com/example/snkrsapp/Data/RemoteData/ProductoDao/TodosProductosRespuesta.kt
@@ -4,7 +4,7 @@ data class TodosProductosRespuesta(
     val idProducto: Int? = null,
     val idMarca: Int = 0,
     val modelo: String = "",
-    val precio: Int? = 0,
+    val precio: Double? = 0.0,
     val talla: Int? = 0,
     val uidVendedor: String? = "",
     val imagenUrl: String? = ""
